Pertemuan 5 Laravel -Show Update and Destroy Data with Postman

+++ database/migrations/2025_05_15_062058_create_books_table.php
@@ -18,8 +18,8 @@ public function up(): void
         $table->integer('price');
         $table->integer('stock');
         $table->string('cover_photo');
-        $table->integer('genre_id');
-        $table->Integer('author_id');
+        $table->bigInteger('genre_id');
+        $table->bigInteger('author_id');
         $table->timestamps();
 
         // // Menambahkan foreign key constraints
